add rubyhash
